Ignore tests that fail when there is a default sysroot built into clang.
A proper solution would have been to modify the tests to handle the default
sysroot correctly.  It is possible, in fact, that the test failures are
legitimate issues and will cause problems later. But, for now, this is the
simplest solution given most of the test involve issues that are not
so interesting.

Disposition: Needs work